Extract shared TextureViewport base for wireframe + PNG viewer
The PNG preview tab (issue #604) had its own hand-rolled viewport, which
duplicated WireframeControl's camera/pan/zoom logic badly enough to
regress five behaviors: grid-colored background, grid rendering,
scrollbars, middle-mouse pan, and wheel zoom-at-cursor.

Fix the duplication at the root: extract TextureViewport (the texture +
pan/zoom camera, per-texture camera cache, grid, palette background,
smooth-zoom animation, pan gesture, wheel zoom-at-cursor, scrollbar
ranges, diagnostics, RenderToBitmap) as a shared base Control. Both
viewers now derive from it:

- WireframeControl : TextureViewport overrides BuildSnapshot to add the
  frame/handle/preview/origin overlay and OnEditPointer* for editing;
  keeps all animation-editor state and services. Public API and every
  Simulate* hook are unchanged, so the existing test suite is the
  regression net (zero test edits).
- PngPreviewControl : TextureViewport is now an empty subclass; inherits
  the full camera/grid/scrollbar/background implementation.

The render-thread snapshot is immutable (TextureViewportSnapshot captured
on the UI thread) with a DrawOverlay delegate so the base draws texture +
grid and the subclass paints its overlay without the base knowing about
frames.

Delete the now-dead PngPreviewScale fit-scale helper + its tests -- the
PNG view uses the shared CanvasTransform.CenterFit path instead.

No behavior change: Core 1518/1518, App 633/633 green; full solution
(incl. Browser/WASM) builds warning-clean.
